Fix opening of multippe context menus
I use `mouseleave` to destroy menu, but the mouse cursor is NOT in menu
initially. So it is possible the following scenario: the user opens
context menu and then optionally move the mouse, but be all time not
inside of menu, and then the opens context menu once more. One more menu
will be opened in the case. The current fix eliminate creating of more
as one menu at the same time.